commiting after long time

+++ server/controllers/branchController.js
@@ -19,7 +19,7 @@ async function addBranch(req,res){
             req.body.latitude,
             req.body.longitude,
             req.body.location_name,
-            req.body.created_user,
+            req.id,
         ]
       const[result]= await db.execute('INSERT INTO branch_locations(latitude,longitude,location_name,created_user,created_time) VALUES (?,?,?,?,?)',[...values,getDateTime()])
       res.send("done add")
@@ -32,6 +32,9 @@ async function addBranch(req,res){
 async function deleteBranch(req,res){
     try {
         const branch_id=req.params.id;
+        const [checkAdmin]= await db.execute('SELECT branch_location_id FROM employees WHERE is_admin="yes"')
+       checkAdmin.sort((a,b)=>a.branch_location_id-b.branch_location_id)
+       if(checkAdmin[0].branch_location_id===checkAdmin[checkAdmin.length-1].branch_location) return res.status(401).json({data:"can not delete a branch that has admin"})
         const[result]= await db.execute('DELETE FROM branch_locations WHERE id=?',[branch_id])
         res.send("done delete")
     } catch (error) {
@@ -43,12 +46,10 @@ async function updateBranch(req, res) {
   try {
     const { id } = req.params;
     const attributes = req.body;
-    const keys = Object.keys(attributes);
-    const values = Object.values(attributes);
-    const placeholders = keys.map((key) => `${key} = ?`).join(", ");
-    const query = `UPDATE branch_location SET ${placeholders} updated_time=? WHERE id = ?`;
-    const params = [...values, getDateTime(), id];
-    const [result] = await db.execute(query, params);
+    const key = Object.keys(attributes);
+    const value = Object.values(attributes);
+    const placeholders = `${key[0]}="${value[0]}", updated_time="${getDateTime()}" ,updated_by=${req.id}`
+    const [result] = await db.execute(`UPDATE branch_locations SET ${placeholders} WHERE id=${id}`);
     res.send(result);
   } catch (error) {
     console.log(error);
